requirements.txt
python-dotenv
fritzconnection
influxdb_client
segno